Added exception handling for Post model

+++ StreamLine/src/main/java/org/example/StreamLine/Service/PostService.java
@@ -1,7 +1,11 @@
 package org.example.StreamLine.Service;
 
+import org.example.StreamLine.Exceptions.PostNotFoundException;
+import org.example.StreamLine.Exceptions.UserNotFoundException;
 import org.example.StreamLine.Model.Post;
+import org.example.StreamLine.Model.User;
 import org.example.StreamLine.Repository.PostRepository;
+import org.example.StreamLine.Repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -13,14 +17,23 @@ import java.util.List;
 public class PostService implements PostServiceInterface {
 
 	private PostRepository postRepository;
+	private UserRepository userRepository;
 
-	public PostService(PostRepository postRepository) {
+	public PostService(PostRepository postRepository, UserRepository userRepository) {
 		this.postRepository = postRepository;
+		this.userRepository = userRepository;
 	}
 	
 	@Override
-	public List<Post> getPostByUser(Integer userId) {
-		return postRepository.findByUserId(userId);
+	public List<Post> getPostByUser(Integer userId) throws UserNotFoundException {
+		User user = userRepository.findById(userId).orElse(null);
+
+		if(user != null) {
+			return postRepository.findByUserId(userId);
+		} else {
+			throw new UserNotFoundException("User with id " + userId + " doesn't exist");
+		}
+
 	}
 
 	@Override
@@ -36,8 +49,14 @@ public class PostService implements PostServiceInterface {
 	}
 
 	@Override
-	public Post getPost(Integer id) {
-		return postRepository.findById(id).orElse(null);
+	public Post getPost(Integer id) throws PostNotFoundException {
+		Post post = postRepository.findById(id).orElse(null);
+
+		if(post != null) {
+			return post;
+		} else {
+			throw new PostNotFoundException("Post with id " + id + " doesn't exist");
+		}
 	}
 
 	@Override
@@ -46,9 +65,17 @@ public class PostService implements PostServiceInterface {
 	}
 
 	@Override
-	public String deletePostById(Integer id) {
-		postRepository.deleteById(id);
-		return "The Post has been deleted";
+	public String deletePostById(Integer id) throws PostNotFoundException {
+		Post post = postRepository.findById(id).orElse(null);
+
+		if(post != null) {
+			postRepository.deleteById(id);
+			return "The Post has been deleted";
+		} else {
+			throw new PostNotFoundException("Post with id " + id + " doesn't exist");
+		}
+
+
 	}
 
 }
